requirements.txt
pathway[xpack-llm-docs]
pdf2img
unstructured
python-dotenv
